requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
langchain-mongodb
openai==1.3.0
PyPDF2==3.0.0
crewai==0.1.21
crewai-tools==0.0.9
pydantic==2.5.2